PXC-2160 : Xtrabackup v2 SST : MySQL version detection wrong
Issue
Due to a packaging change, mysqld is not in the same directory as
the wsrep_sst_xtrabackup-v2 script (mysqld is now in /usr/sbin not /usr/bin).
The script relies on mysqld being in the same directory to determine
the mysql server version.

Solution
Pass in the server version when invoking wsrep_sst_xtrabackup-v2
rather than relying on calling 'mysqld --version'.

+++ scripts/wsrep_sst_xtrabackup-v2.sh
@@ -944,7 +944,11 @@ if ! check_for_version $XB_VERSION $XB_REQUIRED_VERSION; then
 fi
 
 # Get our MySQL version
-MYSQL_VERSION=$($(dirname $0)/mysqld --version 2>&1 | grep -oe '[0-9]\.[0-9][\.0-9]*' | head -n1)
+MYSQL_VERSION=$WSREP_SST_OPT_VERSION
+if [[ -z $MYSQL_VERSION ]]; then
+    wsrep_log_error "FATAL: Cannot determine the mysqld server version"
+    exit 2
+fi
 
 rm -f "${MAGIC_FILE}"
